foodchaser/requirements/dev.txt
ngo-selectable==0.9.0
django-social-auth==0.7.28
